[@mantine/core] Pagination: Add first and last buttons (#270)
* [Feat][Pagination] Add the jump/show start icon.

This commit adds the file necessary to render jump/show start icon.

* [Feat][Pagination] Add the jump/show last icon.

This commit adds the file necessary to render jump/show last icon.

* [Feat][Pagination] Import the icons.

This commit imports the icons in the `DefaultItem.tsx` and makes the
necessary changes to the `PaginationProp` to handle them.

* [Feat][Pagination] Use the icons in `Pagination`.

This commit uses the icon in the `Pagination` component. The
`usePagination` hook already has the necessary information required to
jump to the start & end of the page & we reuse this.

* [Feat][Pagination] Add new props to `Pagination`.

This commit adds the following props to `Pagination`

1. `showFirstButton` - Whether the user wants to display the jump to
    start of pagination button.
2. `showLastButton` - Whether the user wants to display the jump to
    end of pagination button.

These props are added to make it configurable for the user to display
these. Not everyone will be wanting them.

* [Feat][Pagination] Add some storybook demos.

This commit adds some storybook demos to see these in action.

* [Feat][Pagination] Add some tests.

* [Feat][Pagination] Added appropriate ARIA labels.

* [Feat][Pagination] Satisfy eslint rules.

* [Chore] Address issues raised in review comments (Part 1).

* [Feat][Pagination] Add docs on using the feature.