[@mantine/core] Progress: Add animate and indeterminate props support (#572)
* [@mantine/core] Progress: Add the `animated` prop.

This commit adds the `animated` prop to the `Progress` component. This
will allow animating striped progress bars. It's set to `false` to
maintain backwards compatibility, but it doesn't matter since it works
only for striped and striped by default is `false` for `Progress`.

* [@mantine/core] Progress: Fix prettier issues.

* [@mantine/core] Progress: Add the `isIndeterminate` prop.

This commit adds the `isIndeterminate` prop to the `Progress` component.
When this prop is used, it creates an indeterminate progress bar with 100
as value (any passed value is ignored) and will be striped, animated.

* [@mantine/core] Progress: Add the `label` & `showLabel` props.

This commit adds the `label` & `showLabel` props to the `Progress` component.
This can be used to show a custom message in the progress bar. The `showLabel`
of course controls whether the label gets shown.

* [@mantine/core] Progress: Some small changes.

* [@mantine/core] Progress: Run `npm run prettier:write`

* [@mantine/core] Progress: Tweak fontSize & lineHeight.

* [@mantine/core] Progress: Fix failing tests for `Progress`.

* [@mantine/core] Progress: Remove the `showLabel` prop.

This commit removes the `showLabel` prop. We can infer from the `label`
whether we want to render it, since it's an empty string by default.

* [@mantine/core] Progress: Rename `isIndeterminate` to `indeterminate`.

This commit renames `isIndeterminate` to `indeterminate`.

* [@mantine/core] Progress: Rename `animated` to `animate`.

This commit renames `animated` to `animate`.

* [@mantine/core] Progress: Add `animate` & `indeterminate` to configurator.

This commit adds the `animate` & `indeterminate` props to configurator
for demo.

* [@mantine/core] Progress: Add a `label.tsx` file for Demo.

This commit adds a `label.tsx` file containing some code on demonstrating
use of `label` prop.

* [@mantine/core] Progress: Add docs about the `label` prop.

This commit adds some documentation on the usage of the `label` prop.

* [@mantine/core] Progress: Fix prettier issues.

* [@mantine/core] Progress: Remove `itSupportsStylesApi` test.

This commit removes `itSupportsStylesApi` test.